[re-structure] Rename pakage & files to 'check_if_AN_array_is_sorted_and_rotated'
