+++ arrays_sorting_searching/check_if_an_array_is_sorted_and_rotated/check_if_an_array_is_sorted_and_rotated.go
@@ -1,5 +1,5 @@
 // https://www.geeksforgeeks.org/check-if-an-array-is-sorted-and-rotated/
-package check_if_array_is_sorted_and_rotated
+package check_if_an_array_is_sorted_and_rotated
 
 import (
 	"errors"
